golangci-lint: Disable linter that mandates package comments
We don't typically have package-level comments, even though we arguably
should. And while we don't have any such comments in the general case,
the Revive rule never complained about them. This is because the
reported confidence was setto 0.2, and as a result it never got printed.

This was fixed via Revive v1.2.2 so that these violations are now
reported. Let's thus disable the rule.