proto: Add UserCommitFilesError
Add a new `UserCommitFilesError` message type that will be used to add
structured errors to the `UserCommitFiles` RPC. This necessitates
creating a new `IndexError` error type to enumerates the type of index
operations that may fail.

We limit `IndexError` to a single path despite a potentially large
number of files being modified because we end the RPC on the first
failure, so any subsequent paths that would have failed are not
encountered. `path` is sent as `bytes` in the proto because file paths
are not guaranteed to be valid UTF-8 by many common filesytems, which is
the required encoding for protobuf strings.