operations: Simplify checking if a branch to be created exists
In other RPCs that use updateReferenceWithHooks we need to fetch the ref
to figure out what the commit ID is for locking. In the case of a new
branch we always specify this commit ID as zero. Since we have no commit
ID to pass through, this extra reference check does not give us any
protection. The branch could still be created between the initial
reference check and the actual update.

Removing this extra check means that the error message is already marked
as failed precondition correctly. Although it appears that this error
message is less specific this is still and improvement:

* Having the error marked as a failed precondition means that the error
  does not trickle up to sentry.
* Although the error looks user facing, gitlab-rails ignores the actual
  message.

In future we could capture the actual git errors as part of
updateref.Error. For now the message is hardcoded as part of various
operations endpoints.