housekeeping: Refactor test helpers that create file tree entries
The tests for `CleanStaleData()` use a set of helper functions to create
file tree entries. These functions return an interface type though
instead of the concrete implementation, making them hard to use in tests
that do indeed need the concrete type.

Refactor the helper functions to return concrete types. This can make
them harder to use though given that the `d()` function accepts an array
of the interface, and callers would now potentially have to convert the
concrete-type array to the interface-type array. This is alleviated by
accepting a vararg parameter instead of an array.