localrepo: Make GetReference return the name of a conflicting ref
This error does not only happen when the reference name given is not
specific enough, it can also happen when a ref has a path segments the
same as an existing ref. In this case it is useful to specify at least
one of the conflicting refs to reduce the time required to debug such
errors.