git/stats: Drop `UnpackedObjects()` in favor of `LooseObjects()`
The `UnpackedObjects()` helper function returns the number of loose
objects that are newer than the newest packfile. This behaviour is quite
niche, and indeed there are no users outside of our tests. Furthermore,
even the single test user that doesn't verify the function's behaviour
itself wants to only count the number of loose objects in general.

Remove `UnpackedObjects()` and replace the single test-callsite with
`LooseObjects()`.