Praefect: Replicate to existing repository
Currently when the `ReplicateRepository` RPC is invoked through Praefect
with a target repository that already exists, the RPC will fail with the
error `AlreadyExists`. Since it is perfectly valid to invoke this RPC
with a target repository that already exists, Praefect coordinator logic
was updated to handle routing the RPC whether the target repository
already exists or not.