proto: Reformat Protobuf files
Now that we have fixed our Makefile target to detect out-of-date
Protobuf code, our CI does indeed complain about the generated code. As
it turns out the code is different depending on the Go version that's
used to compile it as Go 1.19 has made changes to gofmt. We're still
using Go 1.18 though.

Reformat the code with Go 1.18 to fix the failing job. While it is
suboptimal to have version-dependent code generation, it is still a lot
better than not having safety-guards around out-of-date code. Engineers
can also work around this by just fetching generated artifacts from the
failing pipeline.