Fix 'sleep' invocation incompatiblity between Linux and macOS
TestAuthBeforeLimit is failing on macOS as the sleep call it is using
in a custom hook is adding the 's' suffix for seconds. 'sleep' on macOS
doesn't support a suffix but expects the input to be seconds. The GNU
version also defaults to seconds but accepts a suffix as well. Fix the
incompatibility by removing the unnecessary suffix.