objectpool: Split up implementation for creating and deleting pools
The "create.go" file holds both the implementations for the `Create()`
and `Delete()` RPCs, which is confusing and makes the latter hard to
find. Split out the latter into a new file "delete.go" to fix this.
While at it, we also move common code that is shared by both RPCs into a
new "util.go" file.