objectpool: Reuse `stats.HasBitmap()` in tests
One of the objectpool tests asserts that we correctly remove bitmaps
from a repository when linking it to a pool. But while we have
`stats.HasBitmap()` to check whether a repository has bitmaps or not,
we manually check the filesystem for any bitmaps.

Convert the test to use `stats.HasBitmap()` to improve reuse of
production-facing code. While at it, get rid of the check whether we can
still list references after having linked the repository to the pool:
this is not really relevant at all to the system under test and thus
essentially pointless.