objectpool: Refactor test to not recreate an empty repo
The testcase to verify that `FetchFromOrigin()` pulls in the expected
set of references is creating its own, empty repository that is to be
used as pool member. This was done because in the past the test setup
always created a seeded repository, whereas the test wanted to have full
control over the repository's contents and thus started with an empty
repository instead. But now that the setup function creates an empty
repository by default this is no longer necessary.

Stop creating a separate repository and just use the one returned by the
test setup. While at it, do some small touchups to the test to make
better use of helper functions.