gittest: Remove now-unused function to fix up test repo
The function `FixGitLabTestRepoForCommitGraphs()` was introduced to fix
our seed repository when used together with commit-graphs, which leads
to errors when a specific deliberately broken commit exists. This helper
thus deleted this commit so that we don't hit the error in our tests.

The last user of this function has now been converted to use unseeded
repositories instead, which means that it is now unused. Remove it.