version: Remove unused module version
We're compiling the Go module version of the Gitaly package into the
code. This was used in the past to locate the correct set of auxiliary
binaries like `gitaly-git2go-v15`. But since we have started to package
those binaries into the main `gitaly` executable we don't need to tell
apart different module versions anymore. We have thus removed the last
user of `version.GetModuleVersion()` in 1ba0ce586 (Hardcode module
version in git2go binary name variable, 2022-06-28).

Drop the infrastructure to expose the module version.