objectpool: Introduce temporary option to use seeded test repos
We're about to convert the objectpool tests to not use seeded test
repositories anymore. Introduce an option for `setupObjectPool()` that
temporarily annotates all callsites that currently depend on the seeded
repository contents so that we can convert them one by one.

The option will be removed again once all callsites have been converted.