tools/protogem: Refactor code to be less dependent on output dir
The `generate-proto-ruby` script assumes a specific output directory
that's always fixed at `ruby/proto`. We're about to drop our in-tree
copy of generated Ruby code in favor of using the `gitaly` Gem, which
will require us to generate the code in a temporary directory.

Refactor the code to be less dependent on the output directory. Like
this we can easily change the location where we generate the code into.
Right now the location is unchanged though, which also demonstrates that
there is no change in the generated code.