tools/protogem: Rename `generate-proto-ruby` to `build-proto-gem`
The `generate-proto-ruby` script is not only generating the Ruby sources
anymore, but also builds the `gitaly` Gem. It wasn't renamed in the
preceding commit though as Git doesn't detect the rename by default,
which would have made the diff hard to read.

Rename the file now to the more fitting `build-proto-gem`.