ci: Trigger QA on merged result
When merged result pipelines are enabled we had to set
GITALY_SERVER_VERSION to the source branch SHA of the MR, otherwise QA
wouldn't find this commit.

This issue was reported in [1], and that's now fixed. So we can now
remove that work around and run QA against merged result commits.

[1]: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6482