proto: Regenerate sources with Go 1.19
The Go sources generated from our Protobuf definitions are dependent on
the Go version as they seem to internally use gofmt. With Go 1.19, gofmt
changed the way that comments are formatted, and this in turn impacts
our generated code.

Now that we have upgraded our CI to run with Go 1.19 by default we need
to regenerate the Go code to match the new formatting. Do so.