catfile: Introduce `ObjectReader`
Now `ObjectReader` will provide a way to run git-cat-file(1) via the
--batch-command mode. This means the new interfaces allow for reading of
both object info and object command via the same interface.

In the following commits we'll replace the usage of the existing
interfaces with the new interface behind a feature flag. Eventually
we'll deprecate/delete the old interfaces.

We also change the tests to directly use the newly introduced interfaces,
while we could do this too behind a feature flag, it would be extra
complexity.

Do note that this means now RequestQueue supports ReadObject and
ReadInfo on the same queue.