internal/errors: New shared validation errors
The new validation errors now added to the shared 'errors' package.
The main idea is to re-use these in the services instead of
creating a new error in place. The 'ErrEmptyStorageName' error
signals that the repository doesn't have 'StorageName' field set to
non-blank value. The 'ErrEmptyRelativePath' does the same, but only
for the 'RelativePath' field.

Part of https://gitlab.com/gitlab-org/gitaly/-/issues/3717