gitaly/service: Shared request validation functionality
A lot of RPCs take a Repository as an input parameter. It is a
required parameter in most cases if not at all. Those we need
to validate it is provided, and it has all required fields set
as well. For that purpose the 'ValidateRepository' function is
declared. It accepts repository entity and checks the repository
is not nil, and it has StorageName and RelativePath fields set and
non-blank. If something is missing the corresponding error will be
returned.

Part of https://gitlab.com/gitlab-org/gitaly/-/issues/3717