git2go: Rename IndexError to UnknownIndexError
Currently we may receive one of six distinct error types from `git2go`
when executing the `UserCommitFiles` RPC, one for each specific failure
we check, such as file not found, and a generic `IndexError` used for
invalid paths and problems returned from `libgit2` itself.

In preparation for converting `UserCommitFiles` to return structured
errors, we will convert all defined error types into a single error
struct with an internal enum specifying the error type. This will most
logically be named `IndexError`, but to use that name we must first
rename the existing `IndexError` type. Given that we return the latter
in cases where an undefined problem has occurred in `libgit2`, let's
rename it to `UnknownIndexError`.