git2go: Add new IndexError with multiple variants
In preparation for converting the `UserCommitFiles` RPC to structured
errors, we need to capture the specific problem that occurred. There are
six error types that we track, and while creating a separate error type
for each of these works fine, it's also verbose and requires significant
boilerplate to convert the error to a proto response.

To mitigate this, create a new `IndexError` type that contains an enum
to specify the type of error. This allows us to cover all six error
types with a single case statement when returning from
`UserCommitFiles`, and centralizes the definition of the potential
errors that can be returned. This is similar to the approach taken with
`updateref.CustomHookError`.

The `Proto()` method will be used in the following commit.