housekeeping: Make `OptimizeRepositoryConfig` public to allow testing it
The `OptimizeRepository()` function accepts a variable number of options
that can be used to tweak its behaviour, which then get applied to the
`optimizeRepositoryConfig`. This config struct is not exposed to callers
outside of the `housekeeping` package. And as it is impossible to peek
into the options without applying them to the config, the end result is
that it's not possible to verify which options are ultimately passed to
`OptimizeRepository()`.

Refactor the code to make the `optimizeRepositoryConfig` public. This
allows us to test options passed to the `housekeeping.Manager` in other
packages.