repository: Introduce strategy for `OptimizeRepository`
The `OptimizeRepository` RPC only supports the "heuristical" strategy to
perform repository optimizations, which optimizes data structures based
on whether or not they are already optimized. While this is great in the
general case where we only want to ensure that a repository is in an
efficient state, it is insufficient in case a caller needs to eagerly
optimize a repository. This may for example be required during incidents
or when a repository got into a weird state.

Introduce a new `Strategy` enum for the `OptimizeRepository` RPC that
lets the caller control the strategy which we pick. Right now, we only
support the "heuristical" strategy.