tests: Make the VCS workaround for unprivileged builds global
The testcfg package has some helper functions that build Go binaries
required by Gitaly at runtime. As Go 1.18 has started to run Git as part
of the build process to embed VCS information into the resulting Git
binaries we had two employ two workarounds:

    - We intercept Git commands and thus had to use a Git command
      factory to set up a proper environment.

    - Git refuses to run in repositories not owned by the current user,
      and because CI runs tests as an unprivileged user this safety
      mechanism kicked in. We thus had to inject the `safe.directory`
      Git configuration into the process.

With the introduction of Go 1.19, we furthermore have to change how we
set up the unprivileged user. As a consequence, we also need to start
building binaries outside of our tests as an unprivileged user and thus
the second issue also hits us during the build process.

Now that we have dropped Go 1.17 though we can just bail and do the easy
thing, which is to pass `-buildvcs=false` to Go and be done. This allows
us to get rid of the nasty build hacks and prepares us for Go 1.19.