ci: Introduce a new "analyze" stage
All jobs that either test or analyze are part of the same "test" stage.
Given that the number of jobs has grown quite large by now this makes it
hard to find the job one is searching for.

Introduce a new "analyze" stage to split up those two groups of jobs.