Expose logs as artifacts when running tests on CI
Gitaly test suite on CI doesn't expose any logs. Sometimes, it's not
easy to re-produce the tests on local environment. This commit exposes
the logs generated while running tests. Those logs are then persisted
as the artifacts of CI jobs for a reasonable amount of time.

The folder storing logs are configured from TEST_LOG_DIR env variable.
The dir is created and injected into testcfg builder and testserver as
the dependency of started Gitaly server.

We can also produce the logs on the local environment with
TEST_LOG_DIR=/tmp/gitaly-logs make test-go

Changelog: other