objectpool: Revert error if object does not exist on fetch
In 3bfe68576 (objectpool: Raise error if object pool does not exist on
fetch, 2022-10-26), we have changed `FetchIntoObjectPool()` to raise an
error if the object pool didn't exist yet instead of silently creating
it. This fixes a bug in Praefect, which needs to know about all RPCs
which create a repository in order to record the new repository.

This change unfortunately breaks some tests in Rails. So let's revert it
for the time being so that we ca first amend these tests.

Changelog: fixed