ci: Install default Go version in macOS job
The macOS job installs an unspecified version of Go that gets determined
by the build image. At the point of writing, this is Go 1.17, which we
are about to drop support for.

Update the build instructions to install the default Go version we're
using in our CI so that we have full control over it. This prepares us
for upgrading to Go 1.18+.