middleware: Construct status based on the error
Each response has a status code. It is based on error returned
from the handler. If no error is returned the status is OK, but
if there is an error entity, it would be used to produce the
correct error code. Usually the status package is used to create
an error with proper code to be returned, but in some cases the
proper wrapping with helper or explicit status creation is not
used. In these cases the Unknown code is returned as a result of
the method invocation. Because the error occurs on the backend
side it should be an Internal code instead.
The existing cancelhandler package renamed and now contains unary
and stream interceptors that are modify returned error according
to the context state and error returned by handler.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/4587