objectpool: Reintroduce error if pool does not exist on fetch
In 3bfe68576 (objectpool: Raise error if object pool does not exist on
fetch, 2022-10-26) we started to raise errors when trying to fetch into
an object pool that doesn't exist. Naturally, tests in Rails dependen on
this behaviour, so we had to revert the change in 5cdd7d49c (objectpool:
Revert error if object does not exist on fetch, 2022-11-02).

Reintroduce the change now that the Rails-side tests have been adjusted
to not depend on this via 5dbda9650f32 (spec/object_pool_service: Don't
assume `#fetch` inits the pool, 2022-11-02).