remote: Fix tests that rely on resolving external URLs
We have recently added a set of tests in 8e8d639d4 (remote: Use
`GetURLAndResolveConfig()`, 2022-09-28) that exercise the new
`ResolveAddress` Protobuf field. These tets rely on resolving specific
URLs, which causes them to hang on my system.

Refactor the tests to run with a "real" HTTP server that we're resolving
the original remote URL to. Like this we avoid hitting any external URLs
and furthermore can verify that the command runs successfully as it has
a proper Git server to clone from.