gittest: Automatically close HTTPServer after test
The `gittest.HTTPServer()` function sets up an HTTP server that is
getting served by git-http-backend(1) so that clients can clone from it.
While this function returns a cleanup function, none of its callers use
it in any special way, but only ever close it in a deferred function
call.

Refactor the code to not return the close function anymore and use
`testing.TB.Cleanup` instead.