linguist: Add benchmark for languageStats
I was curious how much extra time it would take to calculate the
FileCount, so with this change we generate fake language stats with 3
million files of 12 random languages. We write these stats to file and
then measure how long it takes to load them. We compare the times with
and without getting the file count.

    goos: linux
    goarch: amd64
    pkg: gitlab.com/gitlab-org/gitaly/v15/internal/gitaly/linguist
    cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
    BenchmarkLanguageStats
    BenchmarkLanguageStats/totals
    BenchmarkLanguageStats/totals-8         1       4298347856 ns/op
    BenchmarkLanguageStats/allCounts
    BenchmarkLanguageStats/allCounts-8      1       6809731635 ns/op

In this bench it takes about 4 seconds to load the stats from file, when
the FileCount is included it takes about 7 seconds in total.