stats: Implement support for SHA256 object format in push probes
The stats package has a blackbox tester for HTTP-based pushes that allow
the user to observe timings for preconfigured pushes. In order for this
to be configurable, we had to reimplement parts of the Git protocol so
that we can easily drive this.

One part that our implementation does not currently support is the
`object-format` capability that advertises the object format that the
client expects the remote repository to be in. As this capability will
default to "sha1" in case it's not explicitly set, this means that we
don't support SHA256-based pushes.

Add a new configuration that allows the user to specify the object
format that shall be used for the push probe and report the object
format capability depending on that. This is sufficient for enabling
SHA256-based pushes.

Changelog: added