limithandler: Add logging to concurrency limiter
When Limit is called on the concurrency limiter, the caller generally
needs to pass the error through and allow it to bubble up because we
want to maintain the structured gRPC error. Instead of putting the
burden of logging the error on the caller, we can log it from the Limit
function if an unexpected error happens.