backup: Don't override relative paths when creating repos
We're manually overriding relative paths in multiple different tests
when calling to `gittest.CreateRepository()`. The consequence is that we
are testing with a storage layout that does not match what we'd have in
production. This is important in the context of backups though given
that backup locations are derived from the repository's relative path.

Refactor the tests to not override the relative path anymore so that we
instead generate them in the `gittest` package via `NewRepositoryName()`.