backup: Do not create non-bare repository paths for tests
Some of our backup tests create  non-bare repository paths which are
lacking the ".git" suffix in their relative path. This is not what we'd
typically have in production, and the suffix does make a difference in
backup tests given that we explicitly strip it in some cases when we
compute backup paths.

Convert the tests to create bare repository paths.