gittest: Convert `NewRepositoryName()` to always create ".git" suffix
The `NewRepositoryName()` function computes a random repository hash for
the caller. This function has two different modes: one to create a bare
repository path with ".git" suffix, and one non-bare one without it.

There is basically not a single instance though where we'd want this
function to create a repository path without the ".git" suffix. In
production systems we always have that suffix as we only ever have bare
repositories. Which means that tests which do use a non-bare repository
path don't match our typical hashed storage layout. This is something we
should avoid though: it's preferable to have our test data as close to
production data as possible.

In the preceding commits we have already converted all remaining callers
that create non-bare repository paths to stop doing that. Let's thus
drop the parameter and unconditionally create bare repository paths so
that noone can be tempted to create a non-bare one anymore.