catfile: Rename ObjectReader to ObjectContentReader
Rename the file/functions pertaining to ObjectReader to
ObjectContentReader. This is in preparation of the upcoming patches
wherein we will introduce a new ObjectReader which will satisfy both the
interfaces of the current ObjectReader and ObjectInfoReader.

To avoid the clash of naming, we shall rename the current ObjectReader
to ObjectContentReader. Eventually we will delete both
ObjectContentReader and ObjectInfoReader.