catfile: Introduce interface `ObjectQueue`
Introduce a new interface `ObjectQueue` which is the summation of the
two interfaces `ObjectInfoQueue` and `ObjectContentQueue`.

We also modify `ObjectContentReader` and `ObjectInfoReader` to return
this new interface and delete both `ObjectContentQueue` and
`ObjectInfoQueue` as they're no longer used. We need this to ensure that
the new object reader we're going to introduce in the following commits
will satisfy a union of the existing interfaces.