catfile: Introduce interface `ObjectQueue`
Introduce a new interface `ObjectQueue` which is the summation of the
two interfaces `ObjectInfoQueue` and `ObjectContentQueue`. Since we're
moving to using the `--batch-command` flag of `git-cat-file(1)`, it'll
replace the current implementation wherein we use `--batch-check` for
info and `--batch` for content of objects.

Since `--batch-command` will be used for both operations, we provide a
single interface which satisfies both the previous interfaces.

We also modify `ObjectContentReader` and `ObjectInfoReader` to return
this new interface and delete both `ObjectContentQueue` and
`ObjectInfoQueue` as they're no longer used.