catfile: Use `ObjectReader` behind the feature flag
Now that we finally have `ObjectReader` ready to be utilized, we use
this behind the `CatfileBatchCommand` featureflag in `catfile/cache.go`.

We also add the required featuresets for tests and a top level flip of
the featureflag in `testhelper.go`, since this code is used by most of
the RPCs.