localrepo: Use catfile cache in ReadObject
A localrepo already has access to the catfile cache, and ReadCommit()
already uses it to read objects. ReadObject() still was spawning a
git-cat-file(1) process itself.

This change modifies ReadObject() to use the catfile cache. This might
avoid actually spawning a new process. We verify this with the newly
introduced gittest.CountingCommandFactory.

The change is behind a feature flag rolled out through [1].

[1]: https://gitlab.com/gitlab-org/gitaly/-/issues/4662

Milestone: 15.7
Label: maintenance::performance