global: Fix use of different cgroups when serving repos
There are several Git commands required for serving repositories to
clients that typically don't get spawned in the repository itself, but
which get the repository path as argument. These are git-upload-pack(1),
git-upload-archive(1) and git-receive-pack(1), all of which we use in
our `smarthttp` and `ssh` services. And because they don't require us to
set up the repository via the `--git-dir` option, we're currently using
`NewWithoutRepo()` to spawn them.

This has a major downside in the context of cgroups though. In order to
decide the cgroup bucket into which we put spawned commands, we either
hash the repository's information or, if not present, the command line.
As we're not passing a repo when serving them, it means that we always
use the command line for them. This is different from most of the other
RPCs though, where we use the repository information instead. Ultimately
this means that every repository has two different cgroups it frequently
gets assigned to, and that's of course not great.

Fix this bug by switching to `New()` instead of `NewWithoutRepo()`.
While not really necessary, it also shouldn't harm us either to make the
Git directory known to git-upload-pack(1) et al. And given that we now
have a repository context available, it means that we will put serving
commands into the same cgroup as for all the other commands.

Changelog: fixed