operations: Stop using `NewWithDir()` for worktrees
Same as with the preceding commit, we put the Git commands that operate
on worktrees into a different cgroup than Git commands that work on a
bare repository. There is only a single such RPC left though, which is
`UserApplyPatch`.

Convert this RPC to use a new Git option `WithWorktree()`. If set, this
option causes us to execute Git commands with the `-C` switch instead of
with `--git-dir` so that the command is executed as if it was executed
in the given worktree. Like this, we spawn those Git commands in the
same cgroup as would get picked in most of the other RPCs.

Changelog: fixed